+++ src/search_schemes/generator/h2.h
@@ -4,6 +4,6 @@
 
 namespace search_schemes::generator {
 
-auto h2(int N, int minK, int K) -> Scheme;
+auto h2(size_t N, size_t minK, size_t K) -> Scheme;
 
 }
