+++ src/search_schemes/generator/pigeon.h
@@ -4,7 +4,7 @@
 
 namespace search_schemes::generator {
 
-auto pigeon_trivial(int minK, int K) -> Scheme;
-auto pigeon_opt(int minK, int K) -> Scheme;
+auto pigeon_trivial(size_t minK, size_t K) -> Scheme;
+auto pigeon_opt(size_t minK, size_t K) -> Scheme;
 
 }
